Use JSTL CustomTypeConverter for all type conversions
