+++ pulsar-transformations/src/main/java/com/datastax/oss/pulsar/functions/transforms/model/ComputeField.java
@@ -16,9 +16,9 @@
 package com.datastax.oss.pulsar.functions.transforms.model;
 
 import com.datastax.oss.pulsar.functions.transforms.jstl.JstlEvaluator;
+import java.time.Instant;
 import java.time.LocalDate;
 import java.time.LocalTime;
-import java.time.OffsetDateTime;
 import java.util.Set;
 import javax.el.ELException;
 import lombok.AccessLevel;
@@ -128,7 +128,7 @@ public class ComputeField {
         case TIME:
           return LocalTime.class;
         case DATETIME:
-          return OffsetDateTime.class;
+          return Instant.class;
         case BYTES:
           return byte[].class;
         default:
